Switched to mac environment. Seems that mac has modified all the line endings.
The important changes here are that the "\" slashes in the includes were changed to be mac compatible, and a function was added for getting the current directory.

Build fine on the mac, cool. Next step is reconfiguring control to go through scokets to win for generating bodies.

+++ UI/QtFabric/view/qtfabric.cpp
@@ -463,8 +463,8 @@ void convertJSONGenome(QWebView* view, QString& jsonString)
 	for(int i=0; i < nodeCount; i++)
 	{
 		printf("%d: \n", i);
-		printf("start signals: %f \n", startSignals[i]);
-		printf("real signals: %f \n", neuronSignals[i]);
+		printf("start signals: %f \n", startSignals[i].toDouble());
+		printf("real signals: %f \n", neuronSignals[i].toDouble());
 		printf("registers: %f \n", registers[i]);
 		QString act = activationFunctions[i].toString();
